% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_meas.R
\name{Agg_meas}
\alias{Agg_meas}
\title{Aggregation Measures}
\usage{
Agg_meas(
  data,
  column,
  measure = c("k", "Hoover", "PoulinD", "QPcrowd", "Lloydcrowd"),
  group = NULL,
  conf = 0.95,
  r = 2000,
  limit = 2000
)
}
\description{
Aggregation Measures
}
