% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcaCI.R
\name{bcaCI}
\alias{bcaCI}
\title{Calculate bca bootstrap confidence intervals}
\usage{
bcaCI(data, column, measure, conf = 0.95, r = 2000, group = NULL)
}
\arguments{
\item{data}{A data frame with parasite data}

\item{column}{Character, indicating the column of parasite intensity data where values are >= 0}

\item{measure}{Character, indicating either abundance or intensity to be used for calculation}

\item{conf}{Numerical, indicating the confidence level desired}

\item{r}{Numerical, indicated the number of replicates for the bootstrapping function}

\item{group}{Character, allows for grouping the data together and calculating intervals for each group}
}
\value{
Returns an object in the form of a data frame that includes mean and confidence intervals. If data is grouped using the group argument, a column with your grouping variables will also be returned within this data frame.
}
\description{
This function allows you to calculate bootstrap confidence intervals using the bca method.
}
