% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2samp_t_test.R
\name{boot2samp_t_test}
\alias{boot2samp_t_test}
\title{Bootstrapped 2-sample t-test for means}
\usage{
boot2samp_t_test(data, column, group, alternative = "two.sided", r = 2000)
}
\description{
Takes normal input data and calculates a bootstrap two-sample t-test from that data to test for differences in means
}
