% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpCI.R
\name{cpCI}
\alias{cpCI}
\title{Calculate Clopper-Pearson confidence intervals}
\usage{
cpCI(data, column, conf = 0.95, group = NULL)
}
\arguments{
\item{data}{A data frame with parasite data}

\item{column}{Character, indicating the column consisting of binomial presence data in 1,0 form}

\item{conf}{Numerical, indicating the confidence level desired}

\item{group}{Character, allows for grouping the data together and calculating intervals for each group}
}
\value{
Returns an object in the form of a data frame that includes naive prevalence, and confidence intervals. If group = T, this data frame also includes a column with your original groups.
}
\description{
This function allows you to calculate Clopper-Pearson confidence intervals.
}
