% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishTest.R
\name{FishTest}
\alias{FishTest}
\title{Calculating Fisher's Exact Test}
\usage{
FishTest(
  data,
  column,
  group,
  alternative = "two.sided",
  simulate.p.value = FALSE,
  B = 2000,
  conf.int = TRUE,
  conf.level = 0.95
)
}
\arguments{
\item{data}{A data frame consisting of at least one column of parasite presence data in binomial form (0 or 1) as well as a column of groups to be tested against each other.}

\item{column}{Character, indicating the column of binomial parasite presence values.}

\item{group}{Character, indicating the column of groups wanting to be compared.}

\item{alternative}{Character, indicating the type of test to be conducted within the fisher.test() function. Options are "two.sided", "greater", "less".}

\item{simulate.p.value}{Logical, see fisher.test() function for description of this option.}

\item{B}{Numerical, indicating the number of replicates used in the Monte Carlo test}

\item{conf.int}{Logical, indicating whether or not to calculate the odds ratio confidence interval. Only works for data with 2 groups. Data with more groups will not have this option.}

\item{conf.level}{Numerical, indicating the level of confidence to calculate the confidence interval. Only works for data with 2 groups.}
}
\value{
Returns the list object described within the fisher.test() function help page.
}
\description{
This function takes normal input data and calculates p-values from Fisher's Exact Test on your specified groups. If looking for greater or less hypotheses, then make sure you set your factor levels correctly. This function depends on the fisher.test() function from base R. See that function for some extra information.
}
\examples{
data(sex)

# Two sided test
FishTest(sex, "intensity", group = "sex")

# Greater than
FishTest(sex, "intensity", group = "sex", alternative = "greater")

# Less than
FishTest(sex, "intensity", group = "sex", alternative = "less")


}
