% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_index.R
\name{H_index}
\alias{H_index}
\title{Hoover Index}
\usage{
H_index(data, column, group = NULL, conf.int = FALSE, conf = 0.95, r = 2000)
}
\description{
Calculates the Hoover index to quantify parasite aggregation.
}
