% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_est.R
\name{k_est}
\alias{k_est}
\title{Estimating the Negative Binomial Exponent, k}
\usage{
k_est(data, column, conf.int = NULL, conf = 0.975, rep = NULL, limit = 2000)
}
\description{
Estimating the Negative Binomial Exponent, k
}
