% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevCI.R
\name{prevCI}
\alias{prevCI}
\title{Prevalence Confidence Intervals}
\usage{
prevCI(
  data,
  column,
  measure = c("ClopPear", "Blaker", "Stern"),
  group = NULL,
  conf = 0.95,
  tolerance = 1e-05
)
}
\description{
Prevalence Confidence Intervals
}
