% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPcrowding.R
\name{QPcrowding}
\alias{QPcrowding}
\title{Calculate Mean Crowding and BCa Confidence Intervals}
\usage{
QPcrowding(data, column, r = 5000, conf = 0.95, group = NULL)
}
\arguments{
\item{data}{A \verb{data frame} consisting of at least one column of parasite intensity}

\item{column}{Character, indicating the name of the column of parasite intensity}

\item{r}{Numeric, Number of replicates to use within the BCa bootstrap}

\item{conf}{Numeric, Confidence level for the BCa bootstrap.}

\item{group}{Character, indicating the name of the column that contains different groups within your data. Ex. 'sex', 'fire', 'age'.}
}
\value{
Returns a data frame object with three columns: mean crowding, lower and upper limits of the confidence interval.
}
\description{
Calculates parasite crowding from normal host parasite intensity or abundance data.
}
