% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_data.R
\docType{data}
\name{sex}
\alias{sex}
\title{Simulated Parasite Dataset}
\format{
\subsection{\code{sex}}{

\describe{
\item{sex}{Sex of the host}
\item{intensity}{Parasite intensity values of these hosts}
\item{presence}{Values indicating presence of parasites on hosts}
...
}
}
}
\usage{
sex
}
\description{
A simulated data set of parasite presence and intensity. Data was simulated from my own parasitology data.
}
\keyword{datasets}
