% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stern.R
\name{stern}
\alias{stern}
\title{Stern Confidence Intervals}
\usage{
stern(
  data,
  column,
  group = NULL,
  alternative = "t",
  method = "minlike",
  conf = 0.95
)
}
\description{
Stern Confidence Intervals
}
