% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Wilson}
\alias{Wilson}
\title{Wilson Score Interval}
\usage{
Wilson(
  data,
  column,
  group = NULL,
  alternative = "t",
  conf = 0.95,
  correct = FALSE
)
}
\arguments{
\item{data}{A data frame \verb{('tbl', 'tbl_df', or 'data.frame')} that consists of at least one column of parasite presence datas in binomial form (0 or 1)}

\item{column}{Character, indicating the column of parasite presence values}

\item{group}{Character, indicating which column to group data by.}

\item{alternative}{Character, indicating the type of test passed to prop.test(). Is not important for confidence interval construction.}

\item{conf}{Numerical, indicating coverage of the confidence interval}

\item{correct}{Logical, indicating whether or not to apply continuity correction to Wilson score intervals.}
}
\description{
Wilson Score Interval
}
