% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X2test.R
\name{X2test}
\alias{X2test}
\title{Chi-Squared Tests for Data Frames}
\usage{
X2test(data, column, group, simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{data}{An input data frame consisting of at least binomial (0 or 1) parasitism data and a grouping variable column.}
}
\description{
Takes input data in the form of a data frame, manipulates that data, and then runs the chisq.test() function for specified groups.
}
