% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_meas.R
\name{Agg_meas}
\alias{Agg_meas}
\title{Parasite Aggregation Measures}
\usage{
Agg_meas(
  data,
  column,
  measure = c("k", "Hoover", "PoulinD", "QPcrowd", "Lloydcrowd"),
  group = NULL,
  conf = 0.95,
  r = 2000,
  limit = 2000
)
}
\arguments{
\item{data}{A data frame \verb{('tbl', 'tbl_df', or 'data.frame')} consisting of at least one column of parasite intensity values for each host.}

\item{column}{Character, indicating the column consisting of parasite intensity values}

\item{measure}{Character, indicating which measures to be calculated.}

\item{group}{Character, indicating the column that data should be grouped by}

\item{conf}{Numerical, indicating coverage of confidence intervals for these measures}

\item{r}{Numerical, number of replicates for bootstrap resampling}

\item{limit}{Numerical, passed into theta.ml() as maximum number of iterations for MLE.}
}
\value{
If only a single measure is chosen, a data frame consisting of the estimate and confidence interval bound for the measure is returned. If multiple measures are chosen, returns a named_list of data frames structured as mentioned.
}
\description{
\code{Agg_meas()} calculates multiple different metrics quantifying the aggregation of parasites within a host population.
}
