% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowding.R
\name{crowding}
\alias{crowding}
\title{Calculate Mean Crowding and Confidence Intervals}
\usage{
crowding(data, column, r = 2000, conf = 0.95, group = NULL)
}
\description{
Calculate Mean Crowding and Confidence Intervals
}
